+++ src/Hacks/FreeScroll.cpp
@@ -1,5 +1,3 @@
-#ifdef QOLMOD_FREESCROLL
-
 #include <Geode/Geode.hpp>
 #include <Geode/modify/EditorUI.hpp>
 #include "../Client/Client.h"
@@ -8,18 +6,10 @@ using namespace geode::prelude;
 
 class $modify (EditorUI)
 {
-    struct Fields {
-        Module* mod = nullptr;
-    };
-
-    void constrainGameLayerPosition()
+    void constrainGameLayerPosition(float p0, float p1)
     {
-        if (!m_fields->mod)
-            m_fields->mod = Client::GetModule("free-scroll");
-
-        if (!m_fields->mod->enabled)
-            EditorUI::constrainGameLayerPosition();
+        
     }
-};
 
-#endif
+    QOLMOD_MOD_HOOK("free-scroll", "EditorUI::constrainGameLayerPosition")
+};